Moved applyStateChange function to base core class.
